Debug statements added

+++ src/main.ts
@@ -47,6 +47,9 @@ async function configure_oci_cli(privateKey: crypto.KeyObject, publicKey: crypto
   const ociPrivateKeyFile = path.join(home, 'private_key.pem');
   const ociPublicKeyFile = path.join(home, 'public_key.pem');
   const upstTokenFile = path.join(home, 'session');
+
+  console.debug(`OCI Config Dir: ${ociConfigDir}`);
+
   const ociConfig = `[DEFAULT]
   user=${ociUser}
   fingerprint=${ociFingerprint}
@@ -57,16 +60,23 @@ async function configure_oci_cli(privateKey: crypto.KeyObject, publicKey: crypto
   `;
 
   await io.mkdirP(ociConfigDir);
+  
+  if (!fs.existsSync(ociConfigDir)) {
+    throw new Error('Unable to create OCI Config folder');
+  }
+  console.debug(`Created OCI Config folder: ${ociConfig}`);
 
   fs.writeFileSync(ociConfigFile, ociConfig);
   fs.writeFileSync(
     ociPrivateKeyFile,
     privateKey.export({ type: 'pkcs1', format: 'pem' }) as string
   );
+
   fs.writeFileSync(
     ociPublicKeyFile,
     publicKey.export({ type: 'spki', format: 'pem' }) as string
   );
+
   fs.writeFileSync(upstTokenFile, upstToken);
 }
